Move Natural to separate utility module
